requirements.txt
Django>=4.2,<5.0
djangorestframework>=3.14,<4.0
djangorestframework-simplejwt>=5.3
django-filter>=23.0
drf-spectacular>=0.27
python-decouple>=3.8
psycopg2-binary>=2.9
gunicorn>=21.2
pytest>=7.4
pytest-django>=4.5
pytest-cov>=4.1
factory-boy>=3.3
